requirements.txt
schedule==1.2.0
web3==6.7.0
python-dotenv==1.0.0
pydantic==2.4.2
loguru==0.7.2
retry==0.9.2 